@@ -9,12 +9,13 @@
     "@testing-library/user-event": "^13.5.0",
     "date-fns": "^4.1.0",
     "firebase": "^11.6.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^3.0.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
-    "react-to-pdf": "^2.0.0",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
